pyproject.toml
[project]
name = "transcription"
version = "0.1.0"
requires-python = ">=3.13"
description = "Diarization and transcription of meetings using Whisper and pyannote.audio"
authors = [
    { name = "Pablo Alcaraz" }
]
dependencies = [
    "openai-whisper>=20250625",
    "pyannote-audio>=4.0.4",
]
